refactor(LadderReward): LadderReward 추가
